Initilize video or camera

+++ examples/cuda/FreeType.cpp
@@ -91,12 +91,12 @@ void make_dlist ( FT_Face face, char ch, GLuint list_base, GLuint * tex_base ) {
 	//first we need to move over a little so that
 	//the character has the right amount of space
 	//between it and the one before it.
-	glTranslatef(bitmap_glyph->left,0,0);
+	glTranslatef((float)bitmap_glyph->left,0,0);
 
 	//Now we move down a little in the case that the
 	//bitmap extends past the bottom of the line 
 	//(this is only true for characters like 'g' or 'y'.
-	glTranslatef(0,bitmap_glyph->top-bitmap.rows,0);
+	glTranslatef(0,(float)bitmap_glyph->top-bitmap.rows,0);
 
 	//Now we need to account for the fact that many of
 	//our textures are filled with empty padding space.
@@ -114,13 +114,13 @@ void make_dlist ( FT_Face face, char ch, GLuint list_base, GLuint * tex_base ) {
 	//so we need to link the texture to the quad
 	//so that the result will be properly aligned.
 	glBegin(GL_QUADS);
-	glTexCoord2d(0,0); glVertex2f(0,bitmap.rows);
+	glTexCoord2d(0,0); glVertex2f(0,(float)bitmap.rows);
 	glTexCoord2d(0,y); glVertex2f(0,0);
-	glTexCoord2d(x,y); glVertex2f(bitmap.width,0);
-	glTexCoord2d(x,0); glVertex2f(bitmap.width,bitmap.rows);
+	glTexCoord2d(x,y); glVertex2f((float)bitmap.width,0);
+	glTexCoord2d(x,0); glVertex2f((float)bitmap.width,(float)bitmap.rows);
 	glEnd();
 	glPopMatrix();
-	glTranslatef(face->glyph->advance.x >> 6 ,0,0);
+	glTranslatef(static_cast<float>( face->glyph->advance.x >> 6 ),0,0);
 
 
 	//increment the raster position as if we were a bitmap font.
@@ -137,7 +137,7 @@ void font_data::init(const char * fname, unsigned int h) {
 	//Allocate some memory to store the texture ids.
 	textures = new GLuint[128];
 
-	this->h=h;
+	this->h=(float)h;
 
 	//Create and initilize a freetype font library.
 	FT_Library library;
@@ -279,7 +279,7 @@ void print(const font_data &ft_font, float x, float y, const char *fmt, ...)  {
 	//down by h. This is because when each character is
 	//draw it modifies the current matrix so that the next character
 	//will be drawn immediatly after it.  
-	for(int i=0;i<lines.size();i++) {
+	for(int i=0;i<(int)lines.size();i++) {
 		
 
 		glPushMatrix();
